youtube downloader - application code/requirements.txt
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
ctk==0.1
customtkinter==5.2.2
darkdetect==0.8.0
ffmpeg==1.4
idna==3.10
mutagen==1.47.0
numpy==1.26.4
packaging==24.1
pillow==10.4.0
pycryptodomex==3.21.0
pytesseract==0.3.13
pytube==15.0.0
pytubefix==8.1.1
requests==2.32.3
tk==0.1.0
urllib3==2.2.3
websockets==13.1
youtube-dl==2021.12.17
yt-dlp==2024.10.22
